+++ wyoming_faster_whisper/faster_whisper/__init__.py
@@ -1,1 +1,0 @@
-from .transcribe import WhisperModel