+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.0.0
+
+- Use faster-whisper PyPI package
+- `--model` can now be a HuggingFace model like `Systran/faster-distil-whisper-small.en`
+
 ## 1.1.0
 
 - Fix enum use for Python 3.11+